requirements.txt
timm==0.4.12
objprint==0.2.3
accelerate==0.23.0
torch==2.2.0
tensorboard
easydict
SimpleITK
monai
nibabel
pyyaml
openslide-python
opencv-python
openpyxl
pandas
histolab
scikit-learn